# docs/swagger.yaml
basePath: /
definitions:
  models.ARPRequest:
    properties:
      cid:
        type: string
      status:
        type: string
    type: object
  models.ARPResponse:
    properties:
      branch_code:
        type: string
      center_code:
        type: string
      cid:
        type: string
      client_mobile_no:
        type: string
      client_name:
        type: string
      created_date:
        type: string
      id:
        type: string
      last_updated_date:
        type: string
      type:
        type: string
      unit_code:
        type: string
    type: object
...
  models.AgentDashboardRequest:
    properties:
      branch_desc:
        type: string
    type: object
  models.AgentDashboardResponse:
    properties:
      aap:
        type: string
      agent_assisted_payment:
        type: string
      bill_payment:
        type: string
      bp:
        type: string
      branch_desc:
        type: string
      cash_in:
        type: string
      cash_out:
        type: string
      ci:
        type: string
      client_name:
        type: string
      co:
        type: string
      id:
        type: string
      sum:
        type: string
      sum_income:
        type: string
      total:
        type: string
      total_income:
        type: string
    type: object
...
  models.ClientProfileRequest:
    properties:
      cid:
        type: string
      mobile:
        type: string
      username:
        type: string
    type: object
  models.ClientProfileResponse:
    properties:
      Agent:
        type: string
      Birthday:
        type: string
      Branch:
        type: string
      Center:
        type: string
      Classification:
        type: string
      Enable:
        type: string
      FullName:
        type: string
      Merchant:
        type: string
      Type:
        type: string
      Unit:
        type: string
      account_name:
        type: string
      account_number:
        type: string
      cid:
        type: string
      id:
        type: string
      insti_name:
        type: string
      is_enabled:
        type: string
      mobile:
        type: string
      username:
        type: string
    type: object
...
  models.EditUseddeviceRequest:
    properties:
      get_branch_code:
        type: string
      get_cid:
        type: string
      get_client_name:
        type: string
      get_client_type:
        type: string
      get_device_id:
        type: string
      get_device_model:
        type: string
      get_device_status:
        type: integer
      get_id:
        type: string
      get_mobile_number:
        type: string
    type: object
  models.EditUseddeviceResponse:
    properties:
      update_listuseddevice:
        type: string
    type: object
...
  models.FailedEnrollmentRequest:
    properties:
      account_number:
        type: string
      client_type:
        example: Client Type
        type: string
      created_date:
        example: Activated Date Start
        type: string
      date_of_birth:
        type: string
      error_message:
        type: string
      last_updated_date:
        example: Activated Date End
        type: string
      mobile_number:
        type: string
    type: object
  models.FailedEnrollmentResponse:
    properties:
      account_number:
        type: string
      client_type:
        type: string
      created_date:
        type: string
      date_of_birth:
        type: string
      device_id:
        type: string
      device_model:
        type: string
      error_message:
        type: string
      id:
        type: string
      mobile_number:
        type: string
    type: object
...
  models.ListofAgentRequest:
    properties:
      a.inst_desc:
        type: string
      c.branch_desc:
        type: string
      c.center_desc:
        type: string
      c.cid:
        type: string
      c.date_and_time:
        example: Activated Date Start
        type: string
      c.mobile_no:
        type: string
      c.unit_desc:
        type: string
      q.date_and_time:
        example: Activated Date End
        type: string
    type: object
  models.ListofAgentResponse:
    properties:
      a.inst_desc:
        type: string
      c.branch_desc:
        type: string
      c.center_desc:
        type: string
      c.cid:
        type: string
      c.date_and_time:
        type: string
      c.fullname:
        type: string
      c.mobile_no:
        type: string
      c.unit_desc:
        type: string
      i.user_name:
        type: string
    type: object
...
  models.OperationDashboardRequest:
    properties:
      last_updated_date:
        type: string
      ref_branch:
        type: string
      trans_post_date:
        type: string
    type: object
  models.OperationDashboardResponse:
    properties:
      Member:
        type: string
      NonMember:
        type: string
      agent_income:
        type: string
      bank_income:
        type: string
      number:
        type: string
      trans_amount:
        type: string
      trans_id:
        type: string
      trans_type:
        type: string
    type: object
...
  models.ResponseModel:
    properties:
      data: {}
      message:
        type: string
      retcode:
        type: string
    type: object
...
  models.UsedDeviceRequest:
    properties:
      cid:
        type: string
      client_type:
        example: Client Type
        type: string
      created_date:
        example: Activated Date Start
        type: string
      device_id:
        type: string
      device_status:
        example: Status
        type: string
      last_updated_date:
        example: Activated Date End
        type: string
      mobile_number:
        type: string
    type: object
  models.UsedDeviceResponse:
    properties:
      android_version:
        type: string
      branch_code:
        type: string
      cid:
        type: string
      client_name:
        type: string
      client_type:
        type: string
      created_date:
        type: string
      device_id:
        type: string
      device_model:
        type: string
      device_status:
        type: string
      mobile_number:
        type: string
    type: object
...
host: 10.21.0.66:1111
info:
  contact:
    email: alex.dequilla@fortress-asya.com
    name: FDSAP-A.Dequilla
  description: KPLUS Webtool Swagger
  license: {}
  termsOfService: http://swagger.io/terms/
  title: Webtool
  version: "2.0"
paths:
...
  /edit_useddevice/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: EditUseddevice Input
        in: body
        name: edituseddeviceInput
        required: true
        schema:
          $ref: '#/definitions/models.EditUseddeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EditUseddeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_agentdashboard/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: AgentDashboard Input
        in: body
        name: agentdashboardInput
        required: true
        schema:
          $ref: '#/definitions/models.AgentDashboardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AgentDashboardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
  /get_arp/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: ARP Input
        in: body
        name: arpInput
        required: true
        schema:
          $ref: '#/definitions/models.ARPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ARPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_clientprofile/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: Clientprofile Input
        in: body
        name: clientprofileInput
        required: true
        schema:
          $ref: '#/definitions/models.ClientProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ClientProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_failedenrollment/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: FailedEnrollment Input
        in: body
        name: failedEnrollmentInput
        required: true
        schema:
          $ref: '#/definitions/models.FailedEnrollmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FailedEnrollmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_listofagent/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: ListofAgent Input
        in: body
        name: listofAgentInput
        required: true
        schema:
          $ref: '#/definitions/models.ListofAgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListofAgentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
  /get_listuseddevice/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: UsedDevice Input
        in: body
        name: usedDeviceInput
        required: true
        schema:
          $ref: '#/definitions/models.UsedDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsedDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_operationdashboard/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: OperationDashboard Input
        in: body
        name: operationdashboardInput
        required: true
        schema:
          $ref: '#/definitions/models.OperationDashboardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OperationDashboardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_remittancelog/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: RemittanceLog Input
        in: body
        name: remittanceLogInput
        required: true
        schema:
          $ref: '#/definitions/models.RemittanceLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RemittanceLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_remittancestatus/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RemittanceStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_slfrequest/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: SlfRequest Input
        in: body
        name: slfrequestInput
        required: true
        schema:
          $ref: '#/definitions/models.SlfRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SlfRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_smslog/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: SmsLog Input
        in: body
        name: smsLogInput
        required: true
        schema:
          $ref: '#/definitions/models.SmsLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SmsLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_transactionlog/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: TransLog Input
        in: body
        name: transLogInput
        required: true
        schema:
          $ref: '#/definitions/models.TransLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TransLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_transconfirmation/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: TransConfirmation Input
        in: body
        name: transconfirmationInput
        required: true
        schema:
          $ref: '#/definitions/models.TransConfirmationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TransConfirmationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /select_useddevice/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: AllUseddevice Input
        in: body
        name: allUsedeviceInput
        required: true
        schema:
          $ref: '#/definitions/models.AllUseddeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllUseddeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
swagger: "2.0"